+++ packages/outstatic/src/components/DocumentSettings/index.tsx
@@ -2,9 +2,9 @@ import Accordion from '@/components/Accordion'
 import DateTimePicker from '@/components/DateTimePicker'
 import DeleteDocumentButton from '@/components/DeleteDocumentButton'
 import DocumentSettingsImageSelection from '@/components/DocumentSettingsImageSelection'
-import Input from '@/components/Input'
 import TagInput from '@/components/TagInput'
-import TextArea from '@/components/TextArea'
+import Input from '@/components/ui/input'
+import TextArea from '@/components/ui/text-area'
 import { DocumentContext } from '@/context'
 import {
   CustomFieldArrayValue,
@@ -16,6 +16,8 @@ import { useRouter } from 'next/navigation'
 import { useContext, useState } from 'react'
 import { RegisterOptions, useFormContext } from 'react-hook-form'
 import { slugify } from 'transliteration'
+import { Button } from '../ui/button'
+import { CheckboxWithLabel } from '../ui/checkbox-with-label'
 
 type DocumentSettingsProps = {
   saveFunc: () => void
@@ -26,13 +28,17 @@ type DocumentSettingsProps = {
 }
 
 interface CustomInputProps {
-  type?: 'text' | 'number'
+  type?: 'text' | 'number' | 'checkbox'
   suggestions?: CustomFieldArrayValue[]
   registerOptions?: RegisterOptions
 }
 
 type ComponentType = {
-  component: typeof Input | typeof TextArea | typeof TagInput
+  component:
+    | typeof Input
+    | typeof TextArea
+    | typeof TagInput
+    | typeof CheckboxWithLabel
   props: CustomInputProps
 }
 
@@ -41,6 +47,7 @@ type FieldDataMapType = {
   Text: ComponentType
   Number: ComponentType
   Tags: ComponentType
+  Boolean: ComponentType
 }
 
 const FieldDataMap: FieldDataMapType = {
@@ -52,7 +59,8 @@ const FieldDataMap: FieldDataMapType = {
     props: {
       suggestions: []
     }
-  }
+  },
+  Boolean: { component: CheckboxWithLabel, props: { type: 'checkbox' } }
 }
 
 const DocumentSettings = ({
@@ -74,32 +82,30 @@ const DocumentSettings = ({
   return (
     <>
       <div className="absolute w-full items-center justify-between flex p-4 border-t z-10 bottom-0 bg-white md:hidden">
-        <button
+        <Button
+          variant="ghost"
+          size="icon"
           onClick={() => setIsOpen(!isOpen)}
-          className="ml-1 inline-flex items-center rounded-lg p-2 text-sm text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 md:hidden"
+          className={`stroke-foreground ${isOpen ? 'bg-accent' : ''}`}
         >
           {isOpen ? <PanelRightClose /> : <PanelRight />}
-        </button>
+        </Button>
         <div className="flex flex-end w-full items-center justify-end gap-4">
           <label htmlFor="status" className="sr-only">
             Status
           </label>
-          <select
-            {...register('status', registerOptions)}
-            name="status"
-            id="status"
-            defaultValue={document.status}
-            className="block cursor-pointer appearance-none rounded-lg border border-gray-300 bg-gray-50 p-2 py-2.5 text-sm text-gray-900 outline-none focus:border-blue-500 focus:ring-blue-500"
-          >
-            <option value="draft">Draft</option>
-            <option value="published">Published</option>
-          </select>
-          <button
-            onClick={saveFunc}
-            type="button"
-            disabled={loading || !hasChanges}
-            className="flex rounded-lg border border-gray-600 bg-gray-800 px-5 py-2.5 text-sm font-medium text-white hover:border-gray-600 hover:bg-gray-700 focus:outline-none focus:ring-4 focus:ring-gray-700 disabled:cursor-not-allowed disabled:bg-gray-600"
-          >
+          <Button asChild variant="select">
+            <select
+              {...register('status', registerOptions)}
+              name="status"
+              id="status"
+              defaultValue={document.status}
+            >
+              <option value="draft">Draft</option>
+              <option value="published">Published</option>
+            </select>
+          </Button>
+          <Button onClick={saveFunc} disabled={loading || !hasChanges}>
             {loading ? (
               <>
                 <svg
@@ -127,13 +133,13 @@ const DocumentSettings = ({
             ) : (
               'Save'
             )}
-          </button>
+          </Button>
         </div>
       </div>
       <aside
         className={`${
           isOpen ? 'block absolute' : 'hidden relative'
-        } md:block w-full border-b border-gray-300 bg-white md:w-64 md:flex-none md:flex-col md:flex-wrap md:items-start md:justify-start md:border-b-0 md:border-l py-6 h-full max-h-[calc(100vh-128px)] md:max-h-[calc(100vh-53px)] scrollbar-hide overflow-scroll`}
+        } md:block w-full border-b border-gray-300 bg-white md:w-64 md:flex-none md:flex-col md:flex-wrap md:items-start md:justify-start md:border-b-0 md:border-l pt-6 pb-16 h-full max-h-[calc(100vh-128px)] md:max-h-[calc(100vh-53px)] scrollbar-hide overflow-scroll`}
       >
         <div className="relative w-full items-center justify-between mb-4 flex px-4">
           <DateTimePicker
@@ -150,16 +156,17 @@ const DocumentSettings = ({
           >
             Status
           </label>
-          <select
-            {...register('status', registerOptions)}
-            name="status"
-            id="status"
-            defaultValue={document.status}
-            className="block cursor-pointer appearance-none rounded-lg border border-gray-300 bg-gray-50 p-2 text-sm text-gray-900 outline-none focus:border-blue-500 focus:ring-blue-500"
-          >
-            <option value="draft">Draft</option>
-            <option value="published">Published</option>
-          </select>
+          <Button asChild variant="select">
+            <select
+              {...register('status', registerOptions)}
+              name="status"
+              id="status"
+              defaultValue={document.status}
+            >
+              <option value="draft">Draft</option>
+              <option value="published">Published</option>
+            </select>
+          </Button>
         </div>
         <div
           className={`flex w-full pb-4 px-4 ${
@@ -177,12 +184,7 @@ const DocumentSettings = ({
               className="hover:bg-slate-200 max-h-[2.25rem]"
             />
           )}
-          <button
-            onClick={saveFunc}
-            type="button"
-            disabled={loading || !hasChanges}
-            className="hidden md:flex rounded-lg border border-gray-600 bg-gray-800 px-5 py-2.5 text-sm font-medium text-white hover:border-gray-600 hover:bg-gray-700 focus:outline-none focus:ring-4 focus:ring-gray-700 disabled:cursor-not-allowed disabled:bg-gray-600"
-          >
+          <Button onClick={saveFunc} disabled={loading || !hasChanges}>
             {loading ? (
               <>
                 <svg
@@ -210,7 +212,7 @@ const DocumentSettings = ({
             ) : (
               'Save'
             )}
-          </button>
+          </Button>
         </div>
         <div className="w-full">
           <Accordion title="Author">
@@ -297,7 +299,6 @@ const DocumentSettings = ({
               )
             })}
         </div>
-        <hr className="pb-16" />
       </aside>
     </>
   )